To disable wait_for_services_timeout use -1 not 0
The documentation for wait_for_services_timeout states that
setting it to 0 will disable the waiting.

However, based on the current implementation at
 https://gitlab.com/gitlab-org/gitlab-runner/blob/01e50620fb2e404d076eae19afde77dd192356f2/executors/docker/docker.go#L465-471 
that does not appear to be the case anymore.

This behavior was changed, perhaps 
accidentally, during this refactor commit from 6 
years ago: 
https://gitlab.com/gitlab-org/gitlab-runner/-/commit/b46be266dbb5d1479f55749c6552b0f7d277a84d#7a056729f6a9c4b323621c1ee3ee2894be489989_451_451

In the current implementation 
setting it to 0 will cause it to silently fall 
back to the default of 30 seconds.

This change updates the document to suggest a negative 
number as the method to disable it, which fits 
with the current config check logic.